Move to_coord_t
