+++ src/utils/ddc_helper.hpp
@@ -291,6 +291,16 @@ struct TypeSeqIntersection<
 };
 
 
+template <class ValidIndexSet>
+struct ToCoord;
+
+template <class... Dims>
+struct ToCoord<VectorIndexSet<Dims...>>
+{
+    using type = Coord<Dims...>;
+};
+
+
 } // namespace detail
 
 /**
@@ -325,4 +335,8 @@ using apply_template_to_type_seq_t = typename detail::ApplyTemplateToTypeSeq<Tem
 template <class TypeSeq1, class TypeSeq2>
 using type_seq_intersection_t =
         typename detail::TypeSeqIntersection<TypeSeq1, TypeSeq2, ddc::detail::TypeSeq<>>::type;
+
+/// Get a coordinate from a TypeSeq
+template <class ValidIndexSet>
+using to_coord_t = typename ToCoord<ValidIndexSet>::type;
 } // namespace ddcHelper
